Add missing API documentation.
This started out as an effort to document the mix function, but then I
noticed a lot of other functions that were missing documentation, and
made an overall pass to finish and make it consistent. Before generating
any documentation at all, I had to fix what appears to be a missing
closing paren in palettes/core.clj so the project as a whole would
compile.